Changes to MovieController and creation of VerifyController for Email Verification
