+++ app/Http/Controllers/MovieController.php
@@ -17,30 +17,7 @@ class MovieController extends Controller
         return view('movies.index');
     }
 
-    public function verifyaccount():View{
-        return view('movies.verifyaccount');
-    }
-
-    public function useractivation(Request $request){
-        $getToken=$request->token;
-        $getToken=Verifytoken::where('token',$getToken)->first();
-
-        if($getToken){
-            $getToken->is_activated=1;
-            $getToken->save();
-
-            $user=User::where('email',$getToken->email)->first();
-            $user->is_activated=1;
-            $user->save();
-            $gettingToken=Verifytoken::where('token',$getToken->token)->first();
-            $gettingToken->delete();
-            return redirect()->route('movie')->with('activated','Your account has been activated successfully');
-        }
-        else{
-            return redirect()->route('verifyaccount')->with('error','Invalid token');
-        }
-    }
-
+   
     /**
      * Show the form for creating a new resource.
      */
